Centralize shared Handlebars helpers
